Rearrange Austin's Work
Also changes the split character from ; to |, as per the assignment requirements
