--- songLib/bin/control/songLibLayout.fxml
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--  Made by Eli Thorpe and Austin Suarez -->
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" 
-minWidth="-Infinity" prefHeight="396.0" prefWidth="472.0" 
-xmlns="http://javafx.com/javafx/15.0.1" 
-xmlns:fx="http://javafx.com/fxml/1" 
-fx:controller="control.SongController">
-   <children>
-      <ListView fx:id="listview" layoutX="6.0" layoutY="6.0" prefHeight="382.0" prefWidth="235.0" />
-      <Text layoutX="301.0" layoutY="31.0" strokeType="OUTSIDE" strokeWidth="0.0" 
-      	text="Song Details">
-         <font>
-            <Font name="System Bold" size="18.0" />
-         </font>
-      </Text>
-      <GridPane layoutX="239.0" layoutY="42.0" prefHeight="123.0" prefWidth="219.0">
-        <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" maxWidth="95.0" minWidth="10.0" prefWidth="60.0" />
-          <ColumnConstraints hgrow="SOMETIMES" maxWidth="167.0" minWidth="10.0" prefWidth="159.0" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-        </rowConstraints>
-         <children>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Title" textAlignment="CENTER" 
-            	wrappingWidth="58.13671875" />
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Artist" textAlignment="CENTER" 
-            	wrappingWidth="60.13671875" GridPane.rowIndex="1" />
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Album" textAlignment="CENTER" 
-            	wrappingWidth="59.13671875" GridPane.rowIndex="2" />
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Year" textAlignment="CENTER" 
-            	wrappingWidth="58.13671875" GridPane.rowIndex="3" />
-            <Label fx:id="titleD" prefHeight="25.0" prefWidth="159.0" text=" " textAlignment="CENTER" 
-            	GridPane.columnIndex="1">
-               <font>
-                  <Font size="13.0" />
-               </font>
-            </Label>
-            <Label fx:id="artistD" prefHeight="25.0" prefWidth="159.0" text=" " textAlignment="CENTER" 
-            	GridPane.columnIndex="1" GridPane.rowIndex="1">
-               <font>
-                  <Font size="13.0" />
-               </font>
-            </Label>
-            <Label fx:id="albumD" prefHeight="23.0" prefWidth="159.0" text=" " textAlignment="CENTER" 
-            	GridPane.columnIndex="1" GridPane.rowIndex="2">
-               <font>
-                  <Font size="13.0" />
-               </font>
-            </Label>
-            <Label fx:id="yearD" prefHeight="24.0" prefWidth="159.0" text=" " GridPane.columnIndex="1" 
-            	GridPane.rowIndex="3">
-               <font>
-                  <Font size="13.0" />
-               </font>
-            </Label>
-         </children>
-      </GridPane>
-      <Text layoutX="298.0" layoutY="198.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Song Controls">
-         <font>
-            <Font name="System Bold" size="18.0" />
-         </font>
-      </Text>
-      <GridPane layoutX="239.0" layoutY="216.0" prefHeight="123.0" prefWidth="219.0">
-         <columnConstraints>
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="95.0" minWidth="10.0" prefWidth="60.0" />
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="167.0" minWidth="10.0" prefWidth="159.0" />
-         </columnConstraints>
-         <rowConstraints>
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-         </rowConstraints>
-         <children>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Title" textAlignment="CENTER" 
-            	wrappingWidth="58.13671875" />
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Artist" textAlignment="CENTER" 
-            	wrappingWidth="60.13671875" GridPane.rowIndex="1" />
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Album" textAlignment="CENTER" 
-            	wrappingWidth="59.13671875" GridPane.rowIndex="2" />
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Year" textAlignment="CENTER" 
-            	wrappingWidth="58.13671875" GridPane.rowIndex="3" />
-            <TextField fx:id="titleA" GridPane.columnIndex="1" />
-            <TextField fx:id="artistA" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-            <TextField fx:id="albumA" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-            <TextField fx:id="yearA" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-         </children>
-      </GridPane>
-      <Button fx:id="addB" layoutX="252.0" layoutY="353.0" mnemonicParsing="false" 
-      	onAction="#action" prefHeight="30.0" prefWidth="63.0" text="Add">
-         <font>
-            <Font name="System Bold" size="14.0" />
-         </font>
-      </Button>
-      <Button fx:id="editB" layoutX="326.0" layoutY="353.0" mnemonicParsing="false" 
-      	onAction="#action" prefHeight="30.0" prefWidth="63.0" text="Edit">
-         <font>
-            <Font name="System Bold" size="14.0" />
-         </font>
-      </Button>
-      <Button fx:id="deleteB" layoutX="400.0" layoutY="353.0" mnemonicParsing="false" 
-      	onAction="#action" text="Delete">
-         <font>
-            <Font name="System Bold" size="14.0" />
-         </font>
-      </Button>
-   </children>
-</AnchorPane>
