--- songLib/songs.csv
@@ -1,10 +1,0 @@
-aaa;bbb;;
-aaa;bcc;;
-asdf;qwer;;
-Bye Bye Bye;NSYNC;No Strings Attached;2005
-ccc;ccd;;
-ccc;ddd;;
-Dancing Queen;ABBA;Arrival;1976
-eee;fff;ggg;1234
-I Want It That Way;Backstreet Boys;Millennium;1999
-Ignition (Remix);R. Kelly;Chocolate Factory;2003
