--- src/songLib/SongLib.java
@@ -1,7 +1,0 @@
-package songLib;
-
-public class SongLib {
-	public static void main(String [] args) {
-		
-	}
-}
